% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PsyScoreMenu.R
\name{Puntuación_del_test}
\alias{Puntuación_del_test}
\title{Puntuaciones de test}
\value{
No devuelve nada. Esta función se usa por su efecto colateral de crear una nueva variable en el conjunto de datos activo.
}
\description{
Cálculo de la puntuación del test a partir de las puntuaciones de las preguntas.
}
\details{
Calcula las puntuaciones del test a partir de variables binarias, numéricas discretas o de factores dos niveles.
Para variables de tipo factor, se toma el último nivel como uno y el otro como 0.
Se permiten NAs, pero el resultado será NA.

Para usar esta función desde R-Commander, seleccione del menú "Psicometría" -> "Datos..." -> "Puntuación del test..."
}
