% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PsyScoreMenu.R
\name{test_score}
\alias{test_score}
\alias{PsyScoreMenu}
\alias{Test_score}
\title{Test score}
\usage{
PsyScoreMenu()
}
\value{
It returns nothing. This function is used for its side effect of creating a new column in the active data set.
}
\description{
Test score calculation from question scores.
}
\details{
Computes the test scores from binary, discrete numeric, or two level factor variables.
For factor type variables, the last level is taken as one and the other as 0.
NA's are allowed, but the result will be NA.

To use this function from R-Commander, select from menu "Psychometry" -> "Data..." -> "Test Score..."
}
